Refactor(package): change package name
