+++ src/main/java/subway/view/MainView.java
@@ -2,7 +2,9 @@ package subway.view;
 
 import java.util.Arrays;
 import java.util.List;
-import subway.message.Output;
+import subway.console.Button;
+import subway.console.Input;
+import subway.console.Output;
 
 /**
  * @author yhh1056
